Enhancements

+++ lib/UI/database/main_image_functions.dart
@@ -1,12 +1,13 @@
+import 'dart:io';
 import 'dart:typed_data';
 import 'dart:ui' as ui;
+
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:solid_color_fills/UI/database/commons.dart';
-import 'package:path_provider/path_provider.dart';
-import 'dart:io';
-import 'package:flutter_wallpaper_manager/flutter_wallpaper_manager.dart';
 import 'package:open_file/open_file.dart';
+import 'package:path_provider/path_provider.dart';
+import 'package:solid_color_fills/UI/database/commons.dart';
 
 final imageProvision = FutureProvider<Uint8List>((ref) async {
   Uint8List uint8list;
@@ -39,37 +40,43 @@ final imageSetter = StateProvider<SetImage>((ref) {
   return SetImage(pngBytes);
 });
 
+const MethodChannel _channel = const MethodChannel('setwallpaper');
+
 class SetImage {
   Uint8List pngBytes;
-  SetImage(this.pngBytes);
+  SetImage(this.pngBytes) : assert(pngBytes != null);
 
   Future<bool> setNow({@required int location}) async {
     // Returns true if success.
     try {
-      bool isDone = false;
-
       if (pngBytes != null) {
-        String tempPath = (await getTemporaryDirectory()).path;
-        String filePath = tempPath + '/temp.png';
-        File file = File(filePath);
-        await file.writeAsBytes(pngBytes);
-
         if (location == 4) {
-          await OpenFile.open(filePath);
-        } else {
-          bool result =
-              await WallpaperManager.setWallpaperFromFile(filePath, location);
+          String tempPath = (await getTemporaryDirectory()).path;
+          String filePath = tempPath + '/temp.png';
+          File file = File(filePath);
 
-          if (result)
-            isDone = true;
-          else
-            isDone = false;
+          await file.writeAsBytes(pngBytes);
 
-          await file.delete();
+          if (await file.exists()) {
+            OpenResult openResult = await OpenFile.open(filePath);
+            if (openResult.type == ResultType.done) await file.delete();
+          }
         }
-      }
 
-      return isDone;
+        // This native method is part of wallpaper plugin, see the plugin's source code.
+        else {
+          _channel.invokeMethod(
+            'setWallpaper',
+            {
+              'data': pngBytes,
+              'location': location,
+            },
+          );
+        }
+
+        return true;
+      } else
+        return false;
     } catch (_) {
       return false;
     }
